Addresses comments
Moves the search page structure to pages/search/index.tsx
